+++ src/pretalx/common/models/settings.py
@@ -14,6 +14,8 @@ settings_hierarkey.add_default('show_schedule', 'True', bool)
 settings_hierarkey.add_default('export_html_on_schedule_release', 'True', bool)
 settings_hierarkey.add_default('custom_domain', '', str)
 
+settings_hierarkey.add_default('display_header_pattern', '', str)
+
 settings_hierarkey.add_default('allow_override_votes', 'False', bool)
 settings_hierarkey.add_default('review_min_score', 0, int)
 settings_hierarkey.add_default('review_max_score', 1, int)