+++ src/pretalx/orga/forms/widgets.py
@@ -1,0 +1,5 @@
+from django.forms import RadioSelect
+
+
+class HeaderSelect(RadioSelect):
+    option_template_name = 'orga/widgets/header_option.html'